+++ demos/asset-request-paths/EntryClient.tsx
@@ -1,0 +1,5 @@
+import { mountApp } from '../../src/client';
+import { routes } from './Routes';
+import './index.css';
+
+mountApp('root', routes);